for SCTP BTL, added Mac OS X support for systems using SCTP NKE (Network Kernel Extension)
This commit was SVN r16729.

+++ ompi/mca/btl/sctp/btl_sctp_frag.c
@@ -408,8 +408,8 @@ repeat11:
                         return false;
                     case EFAULT:
                         opal_output( 0, "mca_btl_sctp_frag_send: writev error (%p, %d)\n\t%s(%d)\n",
-                                frag->iov_ptr[0].iov_base, frag->iov_ptr[0].iov_len,
-                                strerror(opal_socket_errno), frag->iov_cnt );
+                                frag->iov_ptr[0].iov_base, (int) frag->iov_ptr[0].iov_len,
+                                strerror(opal_socket_errno), (int) frag->iov_cnt );
                     default:
                         opal_output(0, "mca_btl_sctp_frag_send: writev failed with errno=%d",
                                 opal_socket_errno);
@@ -552,8 +552,8 @@ repeat:
                         return false;
                     case EFAULT:
                         opal_output( 0, "mca_btl_sctp_frag_send: writev error (%p, %d)\n\t%s(%d)\n",
-                                frag->iov_ptr[0].iov_base, frag->iov_ptr[0].iov_len,
-                                strerror(opal_socket_errno), frag->iov_cnt );
+                                frag->iov_ptr[0].iov_base, (int) frag->iov_ptr[0].iov_len,
+                                strerror(opal_socket_errno), (int) frag->iov_cnt );
                     default:
                         opal_output(0, "mca_btl_sctp_frag_send: writev failed with errno=%d",
                                 opal_socket_errno);
